+++ src/@components/@common/MenuBar/ProfileContainer/index.tsx
@@ -1,8 +1,8 @@
 import { useNavigate } from "react-router-dom";
 
 import { ImgDefaultProfile } from "../../../../asset/image";
-import useUserProfile from "../../../../core/api/myPage";
 import { routePaths } from "../../../../core/routes/path";
+import useUserProfile from "../../../MyPage/hooks/useUserProfile";
 import DefaultProfileContainer from "./DefaultProfileContainer";
 import St from "./style";
