+++ src/@components/BookmarkPage/index.tsx
@@ -1,8 +1,8 @@
-import useUserBookmarks from "../../core/api/bookmark";
 import { HeadingTitle } from "../../util/main/headingTitles";
 import Footer from "../@common/Footer";
 import Header from "../@common/Header";
 import HeadingTitleContainer from "../@common/HeadingTitleContainer";
+import { useUserBookmarks } from "./hooks/useUserBookmarks";
 import MyPiickleItem from "./MyPiickleItem";
 import { St } from "./style";
