+++ src/@components/MainPage/MoodPiickle/index.tsx
@@ -1,13 +1,13 @@
 import { useNavigate } from "react-router-dom";
 import { useSetRecoilState } from "recoil";
 
-import { useCategoryLists } from "../../../core/api/main";
 import { sliderIdxState } from "../../../core/atom/slider";
 import { routePaths } from "../../../core/routes/path";
 import { GTM_CLASS_NAME } from "../../../util/const/gtm";
 import { headingTitles } from "../../../util/main/headingTitles";
 import { gridValue } from "../../../util/main/moodPiickle";
 import HeadingTitleContainer from "../../@common/HeadingTitleContainer";
+import { useCategoryLists } from "../hooks/useCategoryLists";
 import St from "./style";
 
 export type moodPiickle = {
