+++ src/@components/MyPage/MyInfo/index.tsx
@@ -1,8 +1,9 @@
 import { IcChangeProfileBtn } from "../../../asset/icon/index";
 import { ImgDefaultProfile } from "../../../asset/image";
-import useUserProfile, { myPageApi } from "../../../core/api/myPage";
+import { myPageApi } from "../../../core/api/myPage";
 import useModal from "../../@common/hooks/useModal";
 import Loading from "../../@common/Loading";
+import useUserProfile from "../hooks/useUserProfile";
 import NicknameModal from "../NicknameModal";
 import { St } from "./style";
